Implement asset and data extension tools for SFMC MCP server. Refactored server to use modular tool registration.
